:art: restructure page object pattern
