+++ pages/index.js
@@ -1,0 +1,5 @@
+import example_page from './pageObjects/example.page';
+
+export default {
+    example_page
+};
