+++ test/stepDefinitions/when.js
@@ -2,7 +2,7 @@
  * When Steps
  */
 
-const examplePage = require('../pages/example.page');
+const examplePage = require('.././example.page');
 
 module.exports = function(){
