+++ test/utilities/container.js
@@ -1,0 +1,2 @@
+import pages from '../../pages/index';
+export default Object.assign({}, pages);
